Dummy app: move stuff to make sidebar usable, add new dummy route to display scrollablePanel component
