+++ addon/components/o-s-s/scrollable-panel.stories.js
@@ -1,5 +1,4 @@
 import { hbs } from 'ember-cli-htmlbars';
-import { action } from '@storybook/addon-actions';
 
 export default {
   title: 'Components/OSS::ScrollablePanel',
