+++ tests/dummy/app/routes/scrollable.ts
@@ -1,0 +1,7 @@
+import Route from '@ember/routing/route';
+
+export default class Scrollable extends Route.extend({
+  // anything which *must* be merged to prototype here
+}) {
+  // normal class body definition here
+}
