+++ tests/dummy/app/templates/scrollable.hbs
@@ -1,0 +1,72 @@
+<div class="fx-col fx-gap-px-6 padding-px-18">
+  <div class="fx-row fx-gap-px-6">
+    <OSS::Button @label="Open modal" {{on "click" this.openModal}} />
+    <OSS::Button @label="Open sidepanel" {{on "click" this.openSidePanel}} />
+  </div>
+
+  <div style="height: 500px; width:100%">
+    <OSS::ScrollablePanel>
+      <div class="fx-col fx-gap-px-6">
+        {{#each this.loop}}
+          <OSS::InputContainer @placeholder="search" @value={{this.shopUrl}} />
+        {{/each}}
+      </div>
+    </OSS::ScrollablePanel>
+  </div>
+</div>
+
+{{#if this.displayModal}}
+  <OSS::ModalDialog
+    @title="Example modal"
+    @subtitle="subtitle"
+    @close={{this.closeModal}}
+    @disableClickOutside={{false}}
+    @size="md"
+  >
+    <:content>
+      <OSS::ScrollablePanel style="height: 280px">
+        <div class="fx-col fx-gap-px-6">
+          {{#each this.loop}}
+            <OSS::InputContainer @placeholder="search" @value={{this.shopUrl}} />
+          {{/each}}
+        </div>
+      </OSS::ScrollablePanel>
+    </:content>
+    <:footer>
+      <div class="fx-1">
+        <OSS::Icon @icon="fa-info" /><a href="">More info</a>
+      </div>
+      <div class="fx-row fx-gap-px-12">
+        <OSS::Button @skin="default" @label="Close" {{on "click" this.closeModal}} />
+        <OSS::Button @skin="secondary" @label="Save" />
+      </div>
+    </:footer>
+  </OSS::ModalDialog>
+{{/if}}
+
+<OSS::SidePanel
+  @onClose={{this.closeSidePanel}}
+  @backdrop={{true}}
+  @visible={{this.displaySidePanel}}
+  @noPadding={{true}}
+>
+  <:header>
+    <OSS::SidePanel::Header
+      @onClose={{this.closeSidePanel}}
+      @onReturn={{this.onReturn}}
+      @title="This is the side panel's header."
+    />
+  </:header>
+  <:main>
+    <OSS::ScrollablePanel>
+      <div class="fx-col fx-gap-px-6">
+        {{#each this.loop}}
+          <OSS::InputContainer @placeholder="search" @value={{this.shopUrl}} />
+        {{/each}}
+      </div>
+    </OSS::ScrollablePanel>
+  </:main>
+  <:footer>
+    <span>This is the side panel's footer.</span>
+  </:footer>
+</OSS::SidePanel>
